requirements.txt
Flask==2.3.2
python-telegram-bot==20.3
mercadopago==2.2.1
requests==2.31.0
Werkzeug==2.3.6